Fix for GP-7
Resuming a simple two-node instance (with NFS/NIS) works. The main
problem seems to be NIS and NFS being confused by the presence of old
configuration files. The problem is, for the most part, solved by
reverting the files to an unconfigured state before stopping the
instances, but getting to that point also required cleaning up the code
that handles the instance stop/start (mostly so we can include commands
that have to run before an instance is stopped)